Imported Debian patch 2:3.6.6-6+deb7u7

+++ source4/torture/raw/unlink.c
@@ -24,7 +24,6 @@
 #include "libcli/raw/raw_proto.h"
 #include "libcli/libcli.h"
 #include "torture/util.h"
-#include "torture/raw/proto.h"
 
 #define CHECK_STATUS(status, correct) do { \
 	if (!NT_STATUS_EQUAL(status, correct)) { \
@@ -46,9 +45,11 @@ static bool test_unlink(struct torture_context *tctx, struct smbcli_state *cli)
 	bool ret = true;
 	const char *fname = BASEDIR "\\test.txt";
 
-	torture_assert(tctx, torture_setup_dir(cli, BASEDIR), "Failed to setup up test directory: " BASEDIR);
+	if (!torture_setup_dir(cli, BASEDIR)) {
+		return false;
+	}
 
-	printf("Trying non-existent file\n");
+	printf("Trying non-existant file\n");
 	io.unlink.in.pattern = fname;
 	io.unlink.in.attrib = 0;
 	status = smb_raw_unlink(cli->tree, &io);
@@ -178,7 +179,12 @@ static bool test_unlink(struct torture_context *tctx, struct smbcli_state *cli)
 	io.unlink.in.pattern = BASEDIR "\\*.tx?";
 	io.unlink.in.attrib = 0;
 	status = smb_raw_unlink(cli->tree, &io);
-	CHECK_STATUS(status, NT_STATUS_OK);
+	if (torture_setting_bool(tctx, "samba3", false)) {
+		CHECK_STATUS(status, NT_STATUS_NO_SUCH_FILE);
+	}
+	else {
+		CHECK_STATUS(status, NT_STATUS_OK);
+	}
 
 	status = smb_raw_unlink(cli->tree, &io);
 	CHECK_STATUS(status, NT_STATUS_NO_SUCH_FILE);
@@ -208,7 +214,9 @@ static bool test_delete_on_close(struct torture_context *tctx,
 	const char *inside = BASEDIR "\\test.dir\\test.txt";
 	union smb_setfileinfo sfinfo;
 
-	torture_assert(tctx, torture_setup_dir(cli, BASEDIR), "Failed to setup up test directory: " BASEDIR);
+	if (!torture_setup_dir(cli, BASEDIR)) {
+		return false;
+	}
 
 	dio.in.path = dname;
 
@@ -453,10 +461,6 @@ static bool oplock_handler_ack_to_none(struct smbcli_transport *transport,
 	sfinfo.disposition_info.in.file.fnum = fnum;
 	sfinfo.disposition_info.in.delete_on_close = 1;
 	req = smb_raw_setfileinfo_send(ud_cli_state->cli1->tree, &sfinfo);
-	if (!req) {
-		torture_comment(ud_cli_state->tctx, "smb_raw_setfileinfo_send "
-			"failed.");
-	}
 
 	smbcli_close(ud_cli_state->cli1->tree, fnum);
 
@@ -477,6 +481,7 @@ static bool test_unlink_defer(struct torture_context *tctx,
 	bool ret = true;
 	union smb_open io;
 	union smb_unlink unl;
+	uint16_t fnum=0;
 	struct unlink_defer_cli_state ud_cli_state = {};
 
 	if (!torture_setup_dir(cli1, BASEDIR)) {
@@ -513,6 +518,7 @@ static bool test_unlink_defer(struct torture_context *tctx,
 
 	status = smb_raw_open(cli1->tree, tctx, &io);
 	CHECK_STATUS(status, NT_STATUS_OK);
+	fnum = io.ntcreatex.out.file.fnum;
 
 	/* cli2: Try to unlink it, but block on the oplock */
 	torture_comment(tctx, "Try an unlink (should defer the open\n");
